Add files via upload
This is the final update prediction taxonomy file according to the email.   Some genera we couldn't find the corresponding names required by MSL39, but instead, we found names from previous versions or names from NCBI. For these genera, we have marked them with (NCBI) or (Old) in the later sections.